requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
fastapi==0.115.12
Flask==3.1.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
lightgbm==4.6.0
MarkupSafe==3.0.2
numpy==2.2.4
pandas==2.2.3
pydantic==2.11.2
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
threadpoolctl==3.6.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
uvicorn==0.34.0
Werkzeug==3.1.3
xgboost==3.0.0
boto3
